+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.1.0
+* Bumped [`@solana/wallet-adapter-wallets`](https://www.npmjs.com/package/@solana/wallet-adapter-wallets) to `0.19.15`
+* Updated generator in `tool` folder
+
 ## 1.0.1
 * Fixed README.md links to new anchor repository